add api information

+++ domain/src/main/java/com/paya/domain/models/remote/PercentRemoteModel.kt
@@ -10,5 +10,5 @@ data class PercentRemoteModel(
 
 data class PercentEntity(
 	val percent: Float,
-	val price: Float
+	val price: Long
 )
